fix doc string, move http.go tests to http_test.go
